Ensure the core is running after requesting a reset.
According to ARM documentation, after a reset is
requested using the SYSRESETREQ bit in the AIRCR register,
the core should keep running in an endless loop, until
the reset actually occurs. It could take several
instructions until a reset takes place.

To ensure the core is running when the debugger requests
a reset, we start the core after requesting a reset. This
should help with reset problems.