Fix memory writes for ARM platforms
When using a Memory AP to write data into memory, we only
wrote the value into the 'DRW' register of the AP.

The problem is that this does not mean that the value will
be actually written into memory. The SWD protocol requires
that the host does continue to clock the SWD clock after
writing data, to ensure that the SWD transfer can be
completed. In the case of the ST-Link, it seems that
the probe itself does not do this, as least in the
way it is used by probe-rs right now. To ensure that
all transfers can be completed, I have added a read
of the RdBuff register after all writes.