Replace owned Strings and Vectors with Cow
The auto-generated code for the target definitions contained a lot
of Strings and Vectors. For the generated targets, all this data is
constant, and does not have to be owned. This commit replaces all
Strings and Vectors with Cow pointers, meaning that they are now all
statically allocated, but can still be changed if necessary.

This also improves the compile time of probe-rs, it takes now about
half as long to compile as before on my machine.