Run `cargo update`
Not sure what's up, but I'm getting a huge diff in Cargo.lock on every
`cargo build` that I run. The file seems to be formatted weirdly for
some reason, and my Cargo attempts to fix it (maybe a Windows line
ending thing?).

In any case, I thought it best to get it out of the way in a separate
commit.