Determine flashable NVM memory regions based on flash algorithms
Previously, we used the `memory_map` entries in the target description
file to determine whether an NVM address was flashable or not. However,
some target CMSIS-Packs did not include ranges for all flashable areas
(e.g. EEPROM), even though a corresponding flash algorithm exists.

However, if we have a flash algorithm, then we can flash this memory
region, even if no corresponding entry in the `memory_map` exists.

This commit updates the flash loader to use only the flash algorithms to
determine whether an NVM address is flashable or not.

(Support for flashing to RAM is not affected.)