Rework JLink SWD sequence
Removed some comments about the format of the SWD sequence not matching
the specification. In reality, the specification is correct and it seems
it may be the JLink probe that shifts the whole SWDIO sequence by one
bit.

There is a phase shift in SWDIO which is actually covered in the
specification when the ownership of the SWDIO line changes from probe to
target or viceversa. This is documented in [1]. The target places a new
value on the SWDIO line on the rising edge of the clock signal.

In addition to the phase shift, it seems that the JLink probe samples
the SWDIO signal on the falling edge, which would explain why the whole
read sequence is shifted by one bit.

[1]: https://developer.arm.com/documentation/dui0499/j/arm-dstream-target-interface-connections/signal-descriptions/serial-wire-debug
[2]: https://devzone.nordicsemi.com/f/nordic-q-a/9141/swd-interface-changing-phase-from-send-to-receive