Lowercase all algo names.
Previously t2rust lowercased algorithm names. In #569 t2rust was removed, and
algo names are no longer lowercased, so they don't match up with the
"flash_algorithms" field in variants.

Additionally, due to #575, this is not detected when building or running with the
broen YAMLs, instead chips appear as if they have no flash algos.

This lowercases all the names. Validating the algo names match is left as future
work.