arm: Keep core status better updated
The status bits used for the core status are sticky, so we need to ensure
that we read the status register when it's expected to change.

If we don't read it, multiple bits will be set which means it's not
possible anymore to find out why the core is halted.