Improve input event handling (fix #207)
If the access to the probe takes a long time, input events are not
processed properly anymore. It seems that the tick events occur
too often in that case, and the actual input is not processed anymore.

By using recv_timeout, we do not need the Tick events, which seems
to solve this issue.