Prefer exact chip name above substring matches
If there are two chip variants where one's name is a prefix of
the other's, without this change it's not possible to reliably
select the shorter one.

Eg. if there are RP2040 and RP2040_SELFDEBUG, `--chip RP2040` may
accidentally select RP2040_SELFDEBUG.

To avoid that, ignore a variant if it's only a substring match and
there was already some match.