Fix LPC55S69 reset sequence
The current custom reset sequence for the LPC55 triggers the
reset and then puts the target into debug mode using the Debug
Mailbox. While this does ensure that the target will be in a
state for debugging it doesn't match the existing behavior of
reset which is supposed to reset and run code. Switch to only
triggering the debug mailbox if the chip fails to respond after
a reset.