Handle register read failures during LPC55xx reset sequence
When resetting the LPC55xx, there may be a delay before registers
are accessible. This is to be expected and may result in errors when
trying to access registers. The current reset code will bubble up
these failures:

Caused by:
    0: Something during the interaction with the core went wrong
    1: A core architecture specific error occurred
    2: Failed to read register DRW at address 0x0000000c
    3: An error specific to the selected architecture occurred
    4: Target device did not respond to request.

Handle these errors as part of the timeout loop instead.