build: removes libusb1-sys vendored as a default feature
This commit removes the use of `libusb1-sys`'s "vendored" feature by
default. In short, this was causing problems in building libusb v0.6.3.
The issue needs to be reported to the upstream dependency, but in the
meantime, it can be fixed by disabling this default feature.

It should be noted that, without this feature, `libusb1-sys` first
checks to see if you have installed `libusb` yourself. If that does not
exist, it continues to try to install the library from source. So, just
to confirm, nothing is lost here—removing this feature only allows
`libusb1-sys` to first check if you've already installed the library.

See this section of the code for more details:

https://github.com/a1ien/rusb/blob/77468bcaba44c2d073c3bc2afd34454b872ade93/libusb1-sys/build.rs#L218