Add MIMXRT1170 target and 11xx debug sequences
Tested on the Cortex M7 on an MIMXRT1170 EVK. I could download a program
to the on-board flash using probe-rs-cli. I could also debug the Cortex
M7 with GDB. I haven't tested anything with the Cortex M4, although the
core is specified in the target definition.

The target definition came from a FLM in the 1176 device family pack.
It went through the same treatments as the 10xx targets; see the commit
message of d5c8e4a9f for details. I added an extra NVM region for the
Cortex M4's FlexSPI1 alias.

I could borrow most of the debug port start-up steps from the LPC55, so
I broke out the routine into its own function. The TODOs in the routine
also apply to the 1170. I don't have hardware to test this change, but
I'm expecting code review to show consistency.