Fix extraneous MSDAP "StoppedEvent" being sent from `probe-rs-debugger` (#1486)
During the 'step' processing, the debugger would send a `StoppedEvent`
with reason `step`, and immediately after, the regular polling operation
would send another `StoppedEvent` with reason `breakpoint`. The result
of this was that the client would duplicate the identical sequence of
`Threads->Stacktrace->Scopes->Variables` requests. This caused
duplicated processing in `probe-rs-debugger`.

---------

Co-authored-by: Dominik Boehi <dboehi@fastmail.com>