fix: Potential endless recursion in cmsis-dap
In response to a SWD Fault response, the cmsis-dap code tried to read the
debug port control register. If that read failed, it would try to read the
same register again, resulting in endless recursion.

Fixed by adding a separate function to read the CTRL register, which does *not* recurse.