Refactor startup code in debugger (#1511)
I've tried to refactor the "startup" code of the debugger to make it a
bit easier to read, the actual functionality should be unchanged.

It's mostly splitting up the `debug_session` function into different
functions, so it's easier to see what's going on, and trying to reduce
the nesting level, by returning early.

@noppej This probably conflicts with #1506, I'm happy to wait until that
one gets merged and will then update this PR.