Add path_fuzzy_eq() and get_language()
Change debug_info path equality to match against the tail of the path.

The GNU C99 compiler sometimes records a units compile directory as relative. In this case, we cannot predictably construct an absolute path for a compile unit.

To overcome this, we will test equality between a 'sometimes absolute path' and the known absolute path, by matching the tail.