add bitbang probe support
This still needs some work with the error message mapping and the probe
selector interface.  It has been tested with the opentitan verilator
bitbang interface, making it all the way up to the chip identification
(aka read out all of the rv debug module), but failed since the
opentitan is not supported yet.