flashing: Select RAM region by fixed algorithm load address
Some targets have constraints on where flash algorithms can be loaded,
such as if the flash algorithm relies on target ROM routines that use
certain RAM regions as scratch space, if the flash algorithm must run from
secure memory (e.g. Arm TrustZone), or if the algorithm just doesn't use
position-independent code for some reason.

The flash algorithm model already supported fixed load addresses, but the
flasher did not take that into account when selecting a RAM region and so
it worked only if the chosen load address happened to be in the very first
RAM region.

Instead we'll now remove from consideration any RAM region that doesn't
contain the required load address, and therefore choose the first region
that contains that address. This will work only if the chosen load address
leaves enough room for our overhead like the stack and header space, but
this is still an improvement over the previous behavior of selecting
the wrong region altogether.