rtt: Allow targets to constrain the ranges of RTT scanning
For targets that have very large amounts of RAM, that have some RAM regions
that don't make sense for RTT, or that have multiple aliases for the same
RAM, the default behavior of scanning the entirety of all RAM regions is
sub-optimal.

For those targets we'll allow them to specify the what subset of RAM to
scan for a RTT block, with the option of specifying no ranges at all if
there's no reasonable way to narrow down the search space.

The _SEGGER_RTT symbol will still always "win" and prevent attempts to
scan, and so applications are in practice still free to place their RTT
header anywhere as long as they include that symbol in the given ELF
executable, even if that symbol points outside of any of the configured
scan regions.