Issue DAPABORT at startup and after receiving WAIT response from CMSIS-DAP
The ADIv5 spec states that DAPABORT should be issued after receiving
WAIT responses over an extended period. Generally a WAIT response from
a CMSIS-DAP probe means that the probe has timed out after receiving
multiple WAITs from the target, so it would be appropriate to issue
a DAPABORT.

Do the same at startup in case the DP was left in a bad state from
a previous tool run.