rtthost: add channel handling thread spawning
This commit introduces 'RttHost' to allow for calling applications
to hook into RTT channels via 'std::sync::mpsc' channels.

Previously, the rtthost binary contained much of the required
setup. That has been moved out of the main binary and into
the library.

This allows for applications using the library to interface
directly with the target RTT without copying code from the
rtthost binary.