Add note about musl-based systems
On musl-based systems you cannot properly build the CLI tools without
specifying the `RUSTFLAGS=-C target-feature=-crt-static` which makes
rustc no longer attempt to compile things statically (which is the
default on musl-based systems). This PR adds a note about it
in the README.md as well as instruction(s) on how to install the
needed dependency on Alpine Linux.