Reintroduce `stack_size` for MIMXRT1170 targets
858a5c25 changed `target-gen` so that it doesn't suggest a target's
stack size. 858a5c25 also removed the `stack_size` parameter from
existing targets. At the moment, the `stack_size` parameter is important
for the MIMXRT1170 targets, and probe-rs can't program the target
without it.

858a5c25 preserved the probe-rs paths that consider a target's stack
size (see `flash_algorithm.rs`), so this commit re-introduces the value
for a target I'm using. Even if we're not suggesting a target's stack
size when we generate the definition, could we still keep the
`stack_size` parameters for targets that need them?