Remove the Cortex M4 core from MIMXRT1170 targets
The current debug sequence doesn't fully support the CM4. Removing it
lets me focus on a solid debug sequence for the Cortex M7. That's coming
in a child commit.

When re-introducing the CM4, the CM7 needs to release the CM4 from reset
(if it's not already released). Otherwise, the access port for the CM4
isn't available, and `debug_core_start` for the CM4 will insta-fail.
It'll also need to re-release the CM4 after a system reset. Unless we
define a new target and sequence, the existing debug sequence should
handle the fact that the user might want to debug single-core MIMXRT1170
variants.