Use system resets for MIMXRT1170 chips
The previous implementation used a vector reset to get the cores out of
the way so we could program the external flash. It was inspired by
pyOCD's approach, and approaches I've seen in NXP's DFPs. It meant we
were no worst than pyOCD, but we also couldn't execute a newly-flashed
program.

This commit revises the MIMXRT1170 debug sequence. We now perform a true
system reset. The new reset approach requires that we re-initialize the
debug port, then establish the reset context as if we were the boot ROM.
We assume that we're resetting into a program available to FlexSPI1.
Our reset will affect the CM4, if it's utilized, but it does nothing to
release the CM4 or make the CM4 debuggable.

I tested this by flashing small programs built from the NXP SDK, as well
as programs written in Rust. I could debug programs through the debug
shell, and I could place breakpoints through the probe-rs VSCode
extension. The Cortex M7 passes the probe-rs smoke tests.