Add internal method to reset debug probe
Some targets will disable or lose their debug probe connection after a
system reset. The MIMXRT1170 chips exhibit this behavior; the boot ROM
will disable debugging while it executes out of reset. Other open
probe-rs PRs show that the TI CC13xx and CC26xx chips have similar
behaviors.

This commit introduces a reinitialize method to the ARM communication
interface. It refactors the module so that initialization and
re-initialization pass through the same code paths.

It might be nice to represent this reinitialization through the type
system. But today, it wouldn't play well with the debug sequences, since
they only have `&mut _` access to the communication interface.
Therefore, `reinitialize` operates in-place, producing an error that can
be propagated from a debug sequence.

In a child commit, I'll use `reinitialize()` within the MIMXRT1170
`reset_system` debug sequence. As of this commit, the `pub(crate)`
method is unused.