Add support for more Renesas RAx series chips
The RA4M1 was already present. Among the new ones, only RA6M5 has been
tested, but it seemed just as simple to import all the others at the
same time.

Note that the chip known colloquially as "RA6M5" is known to Renesas as
"R7FA6M5", and so on; it is these longer names which are used with
probe-rs's "--chip". For example, with the EK-RA6M5 eval board, use
"--chip R7FA6M5BH".