Add support for the LPC54102 series
This commit adds the YAML file generated by the great `target-gen`-tool
provided in this repository (using the following command):
```shell
$ target-gen arm -f LPC54102 probe-rs/targets
```
This is a sub-series of the LPC54xxx microcontrollers: the [datasheet]
groups the LPC54102 chips of this commit together with the LPC54101-
chips, which are different as they don't have a secondary Cortex-M0+-
core. Therefore they have a different CMSIS-pack and should be in a
separate file.

[datasheet]: https://www.nxp.com/docs/en/data-sheet/LPC5410X.pdf