Increase stack sizes for MIMXRT10xx targets
The new stack overflow checks caught a 16 byte overflow in the 1010
flashing algorithm. To avoid this possibility for other users, this
commit doubles stack sizes for most MIMXRT10xx targets. The 1050 already
specified a 4K stack size for the Hyperflash algorithm, so I'm applying
the same size for the QuadSPI algorithm.