probe-rs: add initial support for black magic probe
This adds support for using a Black Magic Probe with probe-rs.

The Black Magic Probe is a USB device that provides a built-in GDB
server. It supports targets that range from 1.8V to 5V. It presents
two serial ports to the user: One is connected to an external cable,
and the other is a virtual GDB server.

The probe has extended the GDB protocol to add support for driving SWD
and JTAG directly. Additionally, various acceleration functions have
been added to do things like ADIv5 operations to reduce traffic overhead.

This patch adds support for bit-banged JTAG and SWD. Additionally, it
adds supprot for using the probe with ARM targets. It has been tested
with both SWD and JTAG targets using accelerated operations.

It does not yet support RISC-V acceleration, though it is posible to
extend the target to support that.

Signed-off-by: Sean Cross <sean@xobs.io>