Add support for MIMXRT118x M33 memory access
The chip requires authentication before the DP can be accessed. This is
not implemented here, so unlocking has to be done with external tools.

The existing strategy for AP discovery (probing every AP until one fails)
does not work with this chip. Attempting to read IDR from AP5 causes the
whole DP to lock down and require re-authentication. The existing strategy
is turned into a default that can be overridden by debug sequences.

The target description is heavily cut down from the target-gen output
to avoid hitting unsupported features like memory access through the AP
for the M7.