Fix RT1180 defmt-rtt decoding
If the memory where the RTT block's name string ends up is not in a
region in of the target description, probe-rs falls back to interpreting
the thing as raw strings.

The ITCM and DTCM are specified at their maximum sizes, and one or both
will be smaller depending on runtime configuration.