Add warning when refusing to read RTT channel name
When using probe-rs to access a target's RTT stream with an incomplete
target description (missing memory map), channel names will not be
read, resulting in all content being presented as raw strings instead
of interpreting e.g. defmt.

If this happens, it is almost certainly by mistake, so issue a warning.