Adding OL23D0 support
This adds `run`/`flash`/`reset` support for NXPs programmable UWB
controller, [NXP Trimension OL23D0](https://www.nxp.com/products/wireless-connectivity/trimension-uwb/trimension-ol23d0-fully-customizable-uwb-controller-for-iot:OL23D0).
Before this the only officially supported tool to program this chip was
the [Lauterbach Trace32](https://www.lauterbach.com). :dollar:

This PR includes both flash running and RAM running, and is used
continuously in our HIL rigs.

The things that this needs to handle vs vanilla `thumbv8` is that the
customer's firmware and debug probe only has access to the non-secure
region and must talk to the secure region (in TrustZone terms) to
perform flashing. This also means that if there is no firmware loaded
it will get stuck in the secure region. In order to get into a state
where the MCU can be flashed one needs to stop on the first instruction
in the non-secure region, even if there is no firmware loaded, which is
handled via breakpoints. Depending on secure region firmware version
this "magic address" is at different locations and this PR handles all
of them.

Finally, the watchdog can cause issues in this process and the
documentation recommend disabling it. This is also added.

**GOTCHA:**
If one tried to read any part of the flash that are not programmed
(down to 256 bytes page size) will cause a secure lock-up of the chip.
This really threw us of in the beginning as we could read RAM but not
flash.

All datasheets for this is sadly behind NDAs, so consider us (Kiteshield
AB) maintainers of this target. :)