Make gdb-server compatible with lldb
lldb requires the "pc" register to be lower-case. A capital "PC" confuses
it to a point where it thinks it is not debugging at all.

This was already the case for RISC-V. For consistency, I feel it's better
to simply lower-case every register.

Confirmed that armv6-m and risc-v debugging now works just fine with lldb (tested
on Raspberry Pi Pico and esp32c6).

Also confirmed that gdb debugging still works fine - register names are now
lower-case of course.