flash algorithm: drop unnecessary ARM instructions
The original flash algo header was introduced [at the start of the project
(`FLASH_BLOB_HEADER`)], likely taken from elsewhere, similarly to `pyocd`.

Going off [the discussion at pyocd], all but the first instruction are
unnecessary. They were originally part of a checksum, but now the flash
algorithms just jump to the start of the header to trigger a breakpoint, to
halt the core when they're done.

This change preserves the 32-bit wide instructions by using two
breakpoints.

[the discussion at pyocd]: https://github.com/pyocd/pyOCD/discussions/1252
[at the start of the project (`FLASH_BLOB_HEADER`)]: https://github.com/bobrippling/probe-rs/commit/ac2676604ab8d9c838995413334a0b3d88f95188