Mark lazy-loaded flash algos as loaded
879c53e wants to prevent us from loading flash algorithms multiple
times.  However, the commit never marked the algos as loaded. This
commit marks the algorithms as loaded if it ever succeeded. I observed
no issues with this commit when testing on an MIMXRT1170.

For MIMXRT1170 users, this removes a warning about "no program in flash"
observed during flashing. As of this commit, the tool stops issuing
subsequent resets that would detect the previously-erased flash.