blackmagic: support connect-under-reset
The `--connect-under-reset` command uses the `swj_pins()` function
which this platform previously didn't support. Since this platform
generally cannot manually set the pins, this returned an error
`CommandNotSupportedByProbe`.

However, it does support setting the reset pin directly.

Allow `swj_pins()` to succeed if and only if the only selected pin
is the nRST pin.

Signed-off-by: Sean Cross <sean@xobs.io>