probe-rs: source_location: make instruction address accessible
This change prepares for proper address alignment on ARM architectures
with the Thumbv2 instruction set: We need to find the nearest line of
code at the start of the disassembly range so that we can identify the
correct instruction alignment.

This is a direct preparation of the following change that fixes
the overall disassembly instruction counting approach.

Signed-off-by: Florian Grandel <fgrandel@code-for-humans.de>