probe-rs-tools: dap_server: debug_adapter: ergonomics
The disassembler code assumes LE encoding, this is made explicit for
better readability.

Also the accessibility of the capstone disassembler instantiation is now
local as this function is not meant to be used from outside the module.

Future re-use seems improbable as it is generally rather uncommon to
disassemble from a different encoding than the actual source code.

Signed-off-by: Florian Grandel <fgrandel@code-for-humans.de>