Add MIMXRT1160 target
The target is close enough to the MIMXRT1170, so we use the existing
debug sequence. The TCM regions in the target represent the standard
boot fuse FlexRAM configuration.

The device family packs distribute a flash programming algorithm for
this target. I couldn't get it working, so I wrote this flash
programming algorithm myself.

This commit lets me debug and flash programs on a MIMXRT1160EVK. I use
programs built from NXP's SDK, and other build in Rust. The smoke tester
also passes.