traits: arm: make `ArmMemoryInterface` more generic
Previously, the `ArmMemoryInterface` trait returned
`ArmCommunicationInterface<Initialized>` for several functions, or otherwise
assumed that the underlying object was that type. This works for probes
where the underlying protocol is handled by probe-rs, but does not work for
more advanced probes where the protocol is handled on the probe itself.

The functions `try_as_parts()` and `get_arm_communication_interface()` have
been removed, and in their place a selection of other functions have been
added.

Among other things, this fixes the BlackMagicProbe so that it now functions
again.

Signed-off-by: Sean Cross <sean@xobs.io>