Ignore NACK response to power-up request
On PSOC 6, the debug port is unresponsive for about 20us while the
device is waking from deep sleep. CMSIS-DAP transports attempt to read
RDBUFF immediately after writing the power-up request to CTRL/STAT,
which causes a spurious error.

If we get a NACK from a power-up request, igoroe the error and perform a
line reset to clear the protocol error. (If something actually went
wrong, we'll get another error or timeout while polling for the power-up
acknowledgement.)