blackmagic: implement multi-drop support
Implement multi-drop support when using the Black Magic Probe. This change
modifies the `debug_port_start()` call to take a `&mut dyn DapAccess` rather
than a concrete `&mut ArmCommunicationInterface<Initialized>`.

Much of the initialization code was copied from the existing
`ArmCommunicationInterface<Initialized>` implementation and adapted to work
with accelerated opterations on the Black Magic Probe.

Changes to the traits comes from the need for `mcx` to know whether it's
talking JTAG or SWD.

Signed-off-by: Sean Cross <sean@xobs.io>