Add program counter sampling support to profiling
Some DWT implementations expose the program counter sample register. We
can use this to observe the PC value during profiling. It's similar to
the naive approach, but we don't need to halt the core to take the
measurement.

This commit adds program counter sampling support (PCSR) to `probe-rs
profile`. Tested on an i.MX RT MCU. When compared to naive, I measure
over 4x more samples during the same duration.