Address PSOC 6 reset quirks
On PSOC 6, a system reset resets the SWD interface as well, causing all
requests to return Nack until we reinitialize the interface.

Additionally, PSOC 6 takes some time to enter a halted state after
resetting with the reset catch bit -- the reset request bit clears, then
some time later the halt bit is set. This causes an error if we assume
the procesor is halted and we're able to read CPU registers.