tms570: add initial support for tms570
Add initial support for the TI TMS570.

This adds support for configuring the ICEPICK that sits in front of the
microcontroller as well as a simple target configuration.

It does not yet support flashing the device, nor does it support swapping
endianness, which is required as this part is a Big Endian device.

Signed-off-by: Sean Cross <sean@xobs.io>