flashing: fix targets with multiple cores and algorithms
For targets such as stm32wl, multiple algorithms are present. The current
code does not know how to deal with this, and prints an error:

    $ probe-rs run --chip STM32WL55JCIx test.elf
    Error: An error with the flashing procedure has occurred.

    Caused by:
        Trying to write flash, but found more than one suitable flash loader algorithim marked as default for
    NvmRegion { name: Some("BANK_1"), range: 134217728..134479872, cores: ["cm4", "cm0p"], is_alias: false,
    access: Some(MemoryAccess { read: true, write: false, execute: true, boot: true }) }.
    $

Algorithms can specify which core they support, but this is not currently
being checked. Add a check that the core name matches a given algorithm.

This patch fixes flashing on STM32WL55.

Signed-off-by: Sean Cross <sean@xobs.io>