cortex-m: do not log errors on lockup.
These error messages are unconditionally visible, and look scary to the user, suggesting
something's wrong that they should fix, while in fact it's perfectly normal to see a locked up
core in some situations (for example when flashing a blank chip). I've seen a few people misled by
this on Matrix already.

IMO it should be higher-level tools (the debugger, `probe-rs` run) that inform the user of lockup,
not the low-level ARM arch code. They know when lockup is expected or not (for example it's expected
before flashing, unexpected after).