rp2040: read back value of rescue dp after writing it
Read back the value of the RESCUE_DP after writing it. This works around an
issue on WCH-Link probes where not doing this causes the probe to lock up.

Additionally, it was discovered that the `debug_port_setup()` implementation
was breaking things on this probe. The comment said "This works around an
issue where the multidrop DP becomes unavailable.", however it didn't say
what the issues were. If the problem resurfaces, this code will be brought
back and the comment will be updated with the actual issue it solves.

Closes #3480.

Signed-off-by: Sean Cross <sean@xobs.io>