Update nusb to 0.2.0
This change updates the nusb dependency to 0.2.0. There were a number of
[significant changes] in that version. This change relies heavily on
using nusb in its "blocking" mode, in accordance with prior behavior.
The most significant change was to the internal logic in `InterfaceExt`,
with the bulk read/write helpers getting a non-trivial rewrite
(Buffer/Endpoint usage with explicit timeouts and cancellations).

This change is otherwise largely mechanical and the public API of the
crate was not changed.

[significant changes]: https://github.com/kevinmehall/nusb/releases/tag/v0.2.0