Wait for cancellation to clear, and actually send ZLPs
Per review comments, this change will not exit early when attempts are
made to write or read zero-length buffers.

This change also makes the reader and writer wait for cancellations to
complete.  Cancellations are performed asynchronously, and keep the
endpoint claim alive; This causes situations where the next read/write
will race with the cancellation.