armv7a/armv8a: reset fast mode and catch debug aborts
If an instruction fails, the current implementation will continue to poll to
retrieve the data. However, since the instruction has failed, this will
never occur and the program will wait indefinitely.

Check for data aborts as part of every loop. Additionally, add a timeout to
prevent indefinite polling.

Additionally, clear the Memory Access mode for ARMv8A, and set EXTDCC mode
to 0 for ARMv7A. This will fix connecting to a core that has been
interrupted while transferring files, for example when hitting Control-C
during a flash upload.