jtag: mark JtagSequence members as pub
Mark the members of the `JtagSequence` struct as `pub`. This struct is only
used in `pub fn JtagAccess::shift_raw_sequence(&mut self, sequence:
JtagSequence);` which is a public function, however this function can't be
called as `JtagSequence` is currently a sealed trait.

Mark the members as `pub` rather than `pub(crate)`, allowing raw JTAG
sequences to be created by users of the `probe-rs` library.