downgrade cidr preamble mismatch warning to info
When a CIDR value doesn't match the expected value, print out an `info!()`
message rather than a `warn!()` message.

This is because devices ship with broken romtables, and there's nothing that
either probe-rs or the user can do about it. This change prevents this
message from showing during the course of normal operations where `WARN`
messages are shown by default.