Add custom algos for select iMXRT MCUs
These four flash programming algorithms are now written in Rust, and they're
available at the source linked in each YAML file. The algorithms can program
flash around 2x to 3x faster than the previous algorithms, which were extracted
from NXP's DFPs (except for the 1160, which was a custom closed source impl).

The main trick is that we're lying to probe-rs about how big a page is, so it
reduces the I/O between the host and the probe/MCU. The algorithm's page
program call will chunk up the input buffer into pages.

Tested on the four NXP EVKs for these MCUs. It should be straightforward to
replace the other algorithms, but I don't have the parts / boards readily
available for testing.

I'm comparing against probe-rs 0.30.0. probe-rs run --chip=mimxrt1010, before
this commit:

      Erasing ✔ 100% [####################]  32.00 KiB @  28.70 KiB/s (took 1s)
  Programming ✔ 100% [####################]  30.00 KiB @   8.69 KiB/s (took 3s)
    Verifying ✔ 100% [####################]  30.00 KiB @  58.97 KiB/s (took 1s)
      Finished in 5.18s

Same MCU and program, with this commit:

      Erasing ✔ 100% [####################]  32.00 KiB @  35.23 KiB/s (took 1s)
  Programming ✔ 100% [####################]  32.00 KiB @  39.96 KiB/s (took 1s)
    Verifying ✔ 100% [####################]  32.00 KiB @  41.76 KiB/s (took 1s)
      Finished in 2.59s

probe-rs run --chip=mimxrt1160, before this commit with a bigger program:

      Erasing ✔ 100% [####################] 112.00 KiB @  44.05 KiB/s (took 3s)
  Programming ✔ 100% [####################] 111.75 KiB @   6.76 KiB/s (took 17s)
    Verifying ✔ 100% [####################] 111.75 KiB @  56.54 KiB/s (took 2s)
      Finished in 21.16s

Same MCU and program with this commit:

      Erasing ✔ 100% [####################] 112.00 KiB @  41.52 KiB/s (took 3s)
  Programming ✔ 100% [####################] 112.00 KiB @  44.81 KiB/s (took 2s)
    Verifying ✔ 100% [####################] 112.00 KiB @  44.01 KiB/s (took 3s)
      Finished in 7.86s